% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-imf-themes.R, R/imf-themes.R
\docType{data}
\name{A4_colors}
\alias{A4_colors}
\title{List of IMF Colors}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.
}
\usage{
A4_colors

A4_colors
}
\description{
Predefined color palette using standard IMF colors.

Predefined color palette using standard IMF colors.
}
\keyword{datasets}
