% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-imf-themes.R, R/imf-themes.R
\docType{data}
\name{A4_colors_bar}
\alias{A4_colors_bar}
\title{List of IMF Colors for Bar Charts}
\format{
An object of class \code{character} of length 11.

An object of class \code{character} of length 11.
}
\usage{
A4_colors_bar

A4_colors_bar
}
\description{
Extended IMF color palette for bar charts, including lighter and darker variations.

Extended IMF color palette for bar charts, including lighter and darker variations.
}
\keyword{datasets}
