% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-idata_metadata.r
\name{get_dimension_env}
\alias{get_dimension_env}
\title{Create an environment mapping dimension labels to codes}
\usage{
get_dimension_env(db, dimension)
}
\arguments{
\item{db}{Full SDMX dataflow string, e.g. "IMF.STA:CPI"}

\item{dimension}{Dimension name, e.g. "TYPE_OF_TRANSFORMATION"}
}
\value{
Environment: names = labels, values = codes
}
\description{
Create an environment mapping dimension labels to codes
}
