% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba-utils.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Function to generate dimensions of a plot based on type.}
\usage{
get_dimensions(type = c("single", "full_panel", "half_panel"))
}
\arguments{
\item{type}{Type of chart.}
}
\description{
Function to generate dimensions of a plot based on type.
}
