% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-idata-main.R, R/idata-doc.R
\docType{data}
\name{idata}
\alias{idata}
\title{A list bundling IMF data functions}
\format{
An object of class \code{list} of length 3.

A list with three sub-lists:
\describe{
\item{retrieval}{Functions to fetch data: \code{imfdata_by_key}, \code{imfdata_by_countries_and_series}}
\item{metadata}{Functions to query metadata: \code{imfdata_show_datasets}, \code{get_dimension_names}, \code{get_dimension_env}}
\item{helpers}{Utility functions: \code{make_key_str}}
}
}
\usage{
idata
}
\description{
This object provides grouped access to IMF data functions:
\itemize{
\item retrieval functions
\item metadata functions
\item utility functions
}

Provides a top-level list of functions for retrieving data, metadata,
and helper utilities from the IMF iData API.
}
\seealso{
\code{\link{imfdata_by_key}}, \code{\link{imfdata_by_countries_and_series}},
\code{\link{imfdata_show_datasets}}, \code{\link{make_key_str}}
}
\keyword{datasets}
