% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr-package.R
\docType{package}
\name{imf.bookr-package}
\alias{imf.bookr}
\alias{imf.bookr-package}
\title{imf.bookr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BasBBakkerIMF/imf.bookr}
  \item Report bugs at \url{https://github.com/BasBBakkerIMF/imf.bookr/issues}
}

}
\author{
\strong{Maintainer}: Bas Bakker \email{bbakker@Imf.org}

}
\keyword{internal}
