% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-idata_retrieval.r
\name{imfdata_show_datasets}
\alias{imfdata_show_datasets}
\title{Show available IMF dataflows (with optional B2C auth)}
\usage{
imfdata_show_datasets(needs_auth = FALSE)
}
\arguments{
\item{needs_auth}{Logical(1). Whether to add an OAuth token header.}
}
\value{
A data.frame of available dataflows.
}
\description{
Show available IMF dataflows (with optional B2C auth)
}
