% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr-package.R
\docType{package}
\name{imfidata-package}
\alias{imfidata}
\alias{imfidata-package}
\title{imfidata: Fetches IMF iData}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BasBBakkerIMF/imfidata}
  \item Report bugs at \url{https://github.com/BasBBakkerIMF/imfidata/issues}
}

}
\author{
\strong{Maintainer}: Bas Bakker \email{bbakker@Imf.org}

}
\keyword{internal}
