% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr-package.R, R/zzz.R
\docType{package}
\name{imfidatar-package}
\alias{imfidatar-package}
\title{imfidatar: Fetches IMF iData}
\description{
Access IMF iData in R.

Tools to fetch and clean IMF SDMX data (optionally with auth), build keys,
and convert period strings to end-of-period dates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BasBBakkerIMF/imfidata}
  \item Report bugs at \url{https://github.com/BasBBakkerIMF/imfidata/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/BasBBakkerIMF/imfidata}
  \item Report bugs at \url{https://github.com/BasBBakkerIMF/imfidata/issues}
}

}
\author{
\strong{Maintainer}: Bas Bakker \email{bbakker@Imf.org}

}
\keyword{internal}
