% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-idata_retrieval.r
\name{make_dataset_env}
\alias{make_dataset_env}
\title{Build an environment mapping Name.en -> "agencyID:id" (alphabetical)}
\usage{
make_dataset_env(needs_auth = FALSE)
}
\arguments{
\item{df}{A data.frame with at least columns \code{id}, \code{agencyID}, and \code{Name.en}}
}
\value{
An environment where each Name.en is a symbol whose value is "agencyID:id"
}
\description{
Build an environment mapping Name.en -> "agencyID:id" (alphabetical)
}
