% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-idata_metadata.r
\name{make_dimension_env}
\alias{make_dimension_env}
\title{Create an environment mapping dimension labels to codes}
\usage{
make_dimension_env(dataset_or_db, dimension)
}
\arguments{
\item{dataset_or_db}{Character(1). Either an IMF dataflow string
(e.g. "IMF.STA:CPI") or a DSD ID
(e.g. "DSD_BOP_SI").}

\item{dimension}{Character(1). Dimension name,
e.g. "TYPE_OF_TRANSFORMATION".}
}
\value{
Environment: names = labels, values = codes
}
\description{
Create an environment mapping dimension labels to codes
}
