% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-idata_retrieval.r
\name{make_key_str}
\alias{make_key_str}
\title{Convert a key list to a key string}
\usage{
make_key_str(key)
}
\arguments{
\item{key}{A named list of key components (countries, series, frequency)}
}
\value{
A single string like "US+DE.6.M"
}
\description{
Convert a key list to a key string
}
