% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr.R
\name{plot_bar_chart}
\alias{plot_bar_chart}
\title{Function to create a bar chart with IMF stylings}
\usage{
plot_bar_chart(
  input_data,
  var_x,
  var_y,
  fill_var = imfblue,
  fill_colors = A4_colors_bar,
  bar_order = NULL,
  bar_sort = c("none", "desc", "asc"),
  bar_group_sort = FALSE,
  bar_position = ggplot2::position_stack(),
  bar_outline_color = "black",
  bar_outline_size = 0.25,
  bar_width = 0.5,
  bar_label = FALSE,
  bar_label_rounding = 1,
  bar_label_position = c("out", "in"),
  bar_label_adjustment = c(NULL, NULL),
  bar_label_color = "black",
  bar_label_size = 4,
  x_label = "",
  y_label = "",
  plot_title = "",
  plot_subtitle = "",
  plot_caption = NULL,
  panel = FALSE,
  width = 9.46,
  height = 6.85,
  legend_title = "",
  legend_title_format = ggplot2::element_blank(),
  legend_position = "bottom",
  legend_direction = "horizontal",
  legend_key_height = unit(0.45, "cm"),
  legend_key_width = unit(1.3, "cm"),
  legend_spacing_x = unit(0.2, "cm"),
  legend_spacing_y = unit(0.25, "cm"),
  plotname = NA,
  charttype = "single",
  dpi = 600
)
}
\arguments{
\item{input_data}{Input data such as a data.frame}

\item{var_x, var_y}{Column names in \code{input_data} for x-axis and y-axis, respectively}

\item{fill_var}{Color to fill bars, defaults to \code{imfblue}}

\item{fill_colors}{Colors to use for fill, defaults to \code{A4_colors_bar}}

\item{bar_order}{Order of bars, defaults to \code{NULL}}

\item{bar_sort}{Sorting order of bars, defaults to \code{c("none", "desc", "asc")}}

\item{bar_group_sort}{Whether to group sort, defaults to \code{FALSE}}

\item{bar_position}{Position of bars, defaults to \code{ggplot2::position_stack()}
can also use ggplot2::position_dodge() to show side-by-side}

\item{bar_outline_color}{Outline color of bars, defaults to `"black"``}

\item{bar_width}{Width of bars, defaults to \code{0.5}}

\item{bar_label}{Whether to label bars with a given column, defaults to \code{FALSE}}

\item{bar_label_rounding}{Rounding of labels, defaults to \code{1}}

\item{bar_label_position}{Label position, defaults to \code{c("out", "in")}}

\item{bar_label_adjustment}{Adjustment of labels, defaults to \code{c(NULL, NULL)}}

\item{bar_label_color}{Label color of bars, defaults to \code{"black"}}

\item{bar_label_size}{Label size of bars, defaults to \code{4}}

\item{x_label}{X-axis label, defaults to empty string.}

\item{y_label}{Y-axis label, defaults to empty string.}

\item{plot_title}{Chart title, defaults to empty string.}

\item{plot_subtitle}{Subtitle for chart, defaults to empty string.}

\item{plot_caption}{Caption for chart, defaults to empty string.}

\item{panel}{Whether the output is a panel, affects style that is applied.}

\item{legend_title}{Legend title, defaults to empty string}

\item{legend_title_format}{Legend title format, defaults to \code{ggplot2::element_blank()}}

\item{legend_position}{Legend position, defaults to \verb{"bottom",}}

\item{legend_direction}{Legend direction, defaults to \verb{"horizontal",}}

\item{legend_key_height}{Legend key height, defaults to \verb{unit(0.45, "cm"),}}

\item{legend_key_width}{Legend key width, defaults to \verb{unit(1.3, "cm"),}}

\item{legend_spacing_x}{Legend spacing for x direction, defaults to \verb{unit(0.2, "cm"),}}

\item{legend_spacing_y}{Legend spacing for y direction, defaults to \verb{unit(0.25, "cm"),}}

\item{plotname}{If specified, the output name of the chart. Default is \code{NA}
in which case chart is created but not saved as file.}

\item{charttype}{Specify from "single" (default), "full_panel","half_panel".
Determines the size of the output file if \code{plotname} is provided.}

\item{dpi}{Resolution to use if \code{plotname} is specified and applicable. Defaults to \code{600}}
}
\description{
Creates a \link{ggplot2} bar chart with IMF stylings. Takes in at minimum a
data.frame or similar and column names for x and y.
See also \code{\link[=plot_line_chart]{plot_line_chart()}} and \code{\link[=plot_scatter_chart]{plot_scatter_chart()}}.
Wrapper function for \code{\link[ggpubr:ggbarplot]{ggpubr::ggbarplot()}}, which is a wrapper for \link{ggplot2}.
}
\seealso{
\code{\link[=plot_line_chart]{plot_line_chart()}}, \code{\link[=plot_scatter_chart]{plot_scatter_chart()}}
}
