% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr.R
\name{plot_line_chart}
\alias{plot_line_chart}
\title{Function to create a line chart with IMF stylings}
\usage{
plot_line_chart(
  input_data,
  var_x,
  var_y,
  line_color = imfblue,
  col_palette = A4_colors,
  line_size = 1,
  line_type = "solid",
  plot_type = "l",
  point_color = line_color,
  point_size = line_size,
  point_shape = 19,
  point_label = NULL,
  point_repel = NULL,
  font_label = c(7, imfgrey),
  font_family = primary_font,
  x_label = "",
  y_label = "",
  plot_title = "",
  plot_subtitle = "",
  plot_caption = NULL,
  panel = FALSE,
  width = 9.46,
  height = 6.85,
  legend_title = "",
  legend_title_format = ggplot2::element_blank(),
  legend_position = "bottom",
  legend_direction = "horizontal",
  legend_key_height = unit(0.45, "cm"),
  legend_key_width = unit(1.3, "cm"),
  legend_spacing_x = unit(0.2, "cm"),
  legend_spacing_y = unit(0.25, "cm"),
  plotname = NA,
  charttype = "single",
  dpi = 600
)
}
\arguments{
\item{input_data}{Input data such as a data.frame}

\item{var_x, var_y}{Column names in \code{input_data} for x-axis and y-axis, respectively}

\item{line_color}{Line and point color, default is imfblue}

\item{col_palette}{Color palette to use when multiple lines are drawn.
Defaults to \link{A4_colors}.}

\item{line_size}{Width of lines, defaults to 2.5}

\item{line_type}{Line type, defaults to "solid", can choose from
"blank", "solid", "dashed", "dotted", "dotdash", "longdash", and "twodash".}

\item{plot_type}{Specify type of chart, default to "l" for line, can also
specify "p" for markers, and "b" for both lines and markers.}

\item{point_color}{Color of markers if \code{plot_type} is "p" or "b", defaults to \code{line_color}.}

\item{point_size}{Color of markers, defaults to \code{line_size}.}

\item{point_shape}{Shape of markers, defaults to 19 (solid circle).}

\item{point_label}{Labels on points.}

\item{point_repel}{Whether to repel labels on points.}

\item{font_label}{Options for fonts of labesl, defaults to c(7, imfgrey).}

\item{font_family}{Font family to use, defaults to Segoe UI if available.}

\item{x_label}{X-axis label, defaults to empty string.}

\item{y_label}{Y-axis label, defaults to empty string.}

\item{plot_title}{Chart title, defaults to empty string.}

\item{plot_subtitle}{Subtitle for chart, defaults to empty string.}

\item{plot_caption}{Caption for chart, defaults to empty string.}

\item{panel}{Whether the output is a panel, affects style that is applied.}

\item{legend_title}{Legend title, defaults to empty string}

\item{legend_title_format}{Legend title format, defaults to \code{ggplot2::element_blank()}}

\item{legend_position, }{Legend position, defaults to \verb{"bottom",}}

\item{legend_direction, }{Legend direction, defaults to \verb{"horizontal",}}

\item{legend_key_height, }{Legend key height, defaults to \verb{unit(0.45, "cm"),}}

\item{legend_key_width, }{Legend key width, defaults to \verb{unit(1.3, "cm"),}}

\item{legend_spacing_x, }{Legend spacing for x direction, defaults to \verb{unit(0.2, "cm"),}}

\item{legend_spacing_y, }{Legend spacing for y direction, defaults to \verb{unit(0.25, "cm"),}}

\item{plotname}{If specified, the output name of the chart. Default is \code{NA}
in which case chart is created but not saved as file.}

\item{charttype}{Specify from "single" (default), "full_panel","half_panel".
Determines the size of the output file if \code{plotname} is provided.}

\item{dpi}{Resolution to use if \code{plotname} is specified and applicable. Defaults to \code{600}}
}
\description{
Create a ggplot2 line chart with IMF stylings. Takes in at minimum a
data.frame or similar and column names for x and y.
See also \code{\link[=plot_bar_chart]{plot_bar_chart()}} and \code{\link[=plot_scatter_chart]{plot_scatter_chart()}}.
Wrapper function for \code{\link[ggpubr:ggline]{ggpubr::ggline()}}, which is a wrapper for \link{pggplot2}.
}
\seealso{
\code{\link[=plot_bar_chart]{plot_bar_chart()}}, \code{\link[=plot_scatter_chart]{plot_scatter_chart()}}
}
