% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookr.R
\name{plot_scatter_chart}
\alias{plot_scatter_chart}
\title{Function to create a scatter chart with IMF stylings}
\usage{
plot_scatter_chart(
  input_data,
  var_x,
  var_y,
  reg_line = TRUE,
  reg_line_color = imfgrey,
  point_color = imfblue,
  point_outline_color = "black",
  point_size = 2.4,
  point_shape = 21,
  point_label = NULL,
  label_repel = TRUE,
  font_label = c(7, imfgrey),
  font_family = primary_font,
  x_label = "",
  y_label = "",
  plot_title = "",
  plot_subtitle = NULL,
  plot_caption = NULL,
  panel = FALSE,
  width = 9.46,
  height = 6.85,
  legend_title = "",
  legend_title_format = ggplot2::element_blank(),
  legend_position = "bottom",
  legend_direction = "horizontal",
  legend_key_height = unit(0.45, "cm"),
  legend_key_width = unit(1.3, "cm"),
  legend_spacing_x = unit(0.2, "cm"),
  legend_spacing_y = unit(0.25, "cm"),
  plotname = NA,
  charttype = "single",
  dpi = 600
)
}
\arguments{
\item{input_data}{Input data such as a data.frame}

\item{var_x, var_y}{Column names in \code{input_data} for x-axis and y-axis, respectively}

\item{reg_line}{Whether to show a regression line, defaults to \code{TRUE}}

\item{reg_line_color}{Color of regression line, defaults to \code{imfgrey}}

\item{point_color}{Color of markers if \code{plot_type} is "p" or "b", defaults to \code{line_color}.}

\item{point_outline_color}{Outline color of markers, defaults to "black"}

\item{point_size}{Marker size, defaults to 2.4}

\item{point_shape}{Marker shae, defaults to 21}

\item{point_label}{Whether to label points, set column name if necessary, defaults to `NULL}

\item{label_repel}{Whether to repel labels, defaults to \code{TRUE}}

\item{font_label}{Font parameters, defaults to \code{c(7, imfgrey)}}

\item{font_family}{Font family, defaults to \code{primary_font} which should be Segoe UI}

\item{x_label}{X-axis label, defaults to empty string.}

\item{y_label}{Y-axis label, defaults to empty string.}

\item{plot_title}{Chart title, defaults to empty string.}

\item{plot_subtitle}{Subtitle for chart, defaults to empty string.}

\item{plot_caption}{Caption for chart, defaults to empty string.}

\item{panel}{Whether the output is a panel, affects style that is applied.}

\item{legend_title}{Legend title, defaults to empty string}

\item{legend_title_format}{Legend title format, defaults to \code{ggplot2::element_blank()}}

\item{legend_position}{Legend position, defaults to \verb{"bottom",}}

\item{legend_direction}{Legend direction, defaults to \verb{"horizontal",}}

\item{legend_key_height}{Legend key height, defaults to \verb{unit(0.45, "cm"),}}

\item{legend_key_width}{Legend key width, defaults to \verb{unit(1.3, "cm"),}}

\item{legend_spacing_x}{Legend spacing for x direction, defaults to \verb{unit(0.2, "cm"),}}

\item{legend_spacing_y}{Legend spacing for y direction, defaults to \verb{unit(0.25, "cm"),}}

\item{plotname}{If specified, the output name of the chart. Default is \code{NA}
in which case chart is created but not saved as file.}

\item{charttype}{Specify from "single" (default), "full_panel","half_panel".
Determines the size of the output file if \code{plotname} is provided.}

\item{dpi}{Resolution to use if \code{plotname} is specified and applicable. Defaults to \code{600}}

\item{fill_var}{If a color is given, use that color to fill markers. If a column name is
given, use that column's values to change marker colors.}

\item{fill_colors}{Colors to use to fill markers, defaults to \code{A4_colors}}
}
\description{
Creates a \link{ggplot2} scatter chart with IMF stylings. Takes in at minimum a
data.frame or similar and column names for x and y.
See also \code{\link[=plot_line_chart]{plot_line_chart()}} and \code{\link[=plot_bar_chart]{plot_bar_chart()}}.
Wrapper function for \code{\link[ggpubr:ggscatter]{ggpubr::ggscatter()}}, which is a wrapper for \link{ggplot2}.
}
\seealso{
\code{\link[=plot_line_chart]{plot_line_chart()}}, \code{\link[=plot_bar_chart]{plot_bar_chart()}}
}
