% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-imf-themes.R, R/imf-themes.R
\name{rgb2}
\alias{rgb2}
\title{Convert RGB Values to Color String}
\usage{
rgb2(red, green, blue)

rgb2(red, green, blue)
}
\arguments{
\item{red}{Red component (0-255).}

\item{green}{Green component (0-255).}

\item{blue}{Blue component (0-255).}
}
\value{
A color string.

A color string.
}
\description{
Converts red, green, and blue values into a color string that can be used in plotting.

Converts red, green, and blue values into a color string that can be used in plotting.
}
