% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ba-utils.R, R/bookr.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot.}
\usage{
save_plot(
  plot,
  filename,
  type = "single",
  dpi = NULL,
  width = NULL,
  height = NULL
)

save_plot(
  plot,
  filename,
  type = "single",
  dpi = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{plot}{The ggplot2 object to be saved.}

\item{filename}{The file name for the saved plot.}

\item{type}{Chart type, defaults to "single".}

\item{dpi}{Dots per inch for output quality (default: IMF standard 600).}

\item{width}{Plot width in inches (default: IMF standard 9.46).}

\item{height}{Plot height in inches (default: IMF standard 6.85).}

\item{plot_name}{Name of file.}
}
\description{
Uses \code{ggsave_imf()} to apply IMF default width, height, and DPI settings.
}
