% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb-imf-themes.R, R/imf-themes.R
\name{scale_imf_colors}
\alias{scale_imf_colors}
\title{Apply IMF Colors to ggplot2 helper function dynamically}
\usage{
scale_imf_colors(n = NULL)

scale_imf_colors(n = NULL)
}
\description{
Apply IMF Colors to ggplot2 helper function dynamically

Apply IMF Colors to ggplot2 helper function dynamically
}
