% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-idata_retrieval.r
\name{imfdata_show_datasets}
\alias{imfdata_show_datasets}
\title{List available IMF datasets (dataflows)}
\usage{
imfdata_show_datasets(needs_auth = FALSE)
}
\arguments{
\item{needs_auth}{Logical: whether to include auth headers}
}
\value{
A data.frame of dataflow metadata
}
\description{
List available IMF datasets (dataflows)
}
